Fix multiple definitions with libstdc++ of GCC.
The libstdc++v3 in GCC reimplements several math
functions if few flags from GLIBC are not provided.
If so, as the libmath already provides these functions,
it leads to multiple definition failure at link time.
Thus, we define these GLIBC flags to make the libstdc++v3 of GCC
to not provide these functions in libstdc++ of the elf, mbr, and cos
toolchain.

ref T11404

Test Plan: newlib, OpenCL

Reviewers: O26 newlib!

Differential Revision: https://phab.kalray.eu/D3176