[kvx] memcmp fix stack corruption if IT user occurs
Summary:
If user interrupt occurs in kvx clusterOS,
the stack of the interrupted context is save on the
user stack itself. As the memcmp asm code doesn't
move the stack pointer, the saved registers may be
corruption if a user interrupt is taken. This commit
moves the stack pointer before using the stack in
memcmp so that if an user interrupt is taken, it
will not be corrupted by interrupt context save.

Test Plan: CI

Reviewers: sjones, amichon, O26 newlib!

Differential Revision: https://phab.kalray.eu/D19276