Rename XOR* mnemonics to EOR*
Summary:
TCA mnemonics all start with an X, XOR was the only exception.  This
renaming makes it so that only TCA mnemonics start by an X.

On KV3-1 and KV3-2, we install synthetics so that XOR still works.
However, the KV4-1 will not expose a XOR mnemonic.

libgloss/ChangeLog:

2023-10-31  Paul Iannetta  <piannetta@kalrayinc.com>

	* kvx-elf/include/kvx/asm_context.h

newlib/ChangeLog:

2023-10-31  Paul Iannetta  <piannetta@kalrayinc.com>

	* libc/machine/kvx/memcmp.S: Rename XOR* mnemonics to EOR*.

Test Plan: CI

Reviewers: O26 newlib!

Differential Revision: https://phab.kalray.eu/D33495